Added refresh token to db. Create login form, sidebar, nav, dashboard in frontend

+++ GroomerManager.Shared/DTOs/Response/LoginResponseDto.cs
@@ -7,7 +7,7 @@ public class LoginResponseDto
     [property: JsonPropertyName("token")]
     public required string Token { get; set; }
     [property: JsonPropertyName("tokenExpired")]
-    public required int TokenExpired { get; set; }
+    public required long TokenExpired { get; set; }
     [property: JsonPropertyName("refreshToken")]
     public required string RefreshToken { get; set; }
 }
